Corrigido Login e Autenticaçao ok
