+++ src/index.html
@@ -14,17 +14,19 @@
   <script src="https://www.gstatic.com/firebasejs/6.2.0/firebase-firestore.js"></script>
   <script type="module" src="./main.js"></script>
   <script>
-  var firebaseConfig = {
-  apiKey: "api-key",
-  authDomain: "rede-social-tech.firebaseapp.com",
-  databaseURL: "https://rede-social-tech.firebaseio.com",
-  projectId: "rede-social-tech",
-  storageBucket: "rede-social-tech.appspot.com",
-  messagingSenderId: "sender-id",
-  appID: "app-id",
-};
-  firebase.initializeApp(firebaseConfig);
-  </script>
-
+    // Your web app's Firebase configuration
+    var firebaseConfig = {
+      apiKey: "AIzaSyDfD8d6v_7hXe4coek1xYO_mCEPyZKUdcI",
+      authDomain: "rede-social-tech.firebaseapp.com",
+      databaseURL: "https://rede-social-tech.firebaseio.com",
+      projectId: "rede-social-tech",
+      storageBucket: "rede-social-tech.appspot.com",
+      messagingSenderId: "417687051409",
+      appId: "1:417687051409:web:8dd4113911ee43a7447330",
+      measurementId: "G-2DZ77BZ3HW"
+    };
+    // Initialize Firebase
+    firebase.initializeApp(firebaseConfig);
+    </script>
 </body>
 </html>
