+++ src/main.js
@@ -5,3 +5,5 @@ function init() {
 }
 
 window.addEventListener('load', init);
+
+window.addEventListener("#Home", Login, false);
