+++ src/pages/signout.js
@@ -1,0 +1,8 @@
+firebase.auth().signOut().then(function() {
+  alert ('Sign-out successful.')
+}).catch(function(error) {
+  // An error happened.
+});
+
+  window.location.href = '#singOut';
+export default SignOut;
